Add server-side configurations and ring enchantment updates
This update introduces server-side configurations for ring enchantments. It also refactors how the "Susurrus" ring enchantment is applied. The ticks are now managed through a new RingComponent instead of via tags on the enchantment item stack. This simplifies the code and improves performance. The targets of enchantments are now triggered based on a configurable server-side setting. This provides a greater level of customization for server admins. Gradle variables are also now used in the pack.mcmeta and fabric.mod.json

+++ src/main/java/org/ecorous/dwarventreasures/world/item/enchantment/DwarvenTreasuresEnchantments.java
@@ -1,5 +1,6 @@
 package org.ecorous.dwarventreasures.world.item.enchantment;
 
+import dev.emi.trinkets.api.TrinketsApi;
 import net.minecraft.core.BlockPos;
 import net.minecraft.core.Registry;
 import net.minecraft.core.particles.ParticleTypes;
@@ -7,6 +8,7 @@ import net.minecraft.core.registries.BuiltInRegistries;
 import net.minecraft.tags.EntityTypeTags;
 import net.minecraft.util.Mth;
 import net.minecraft.util.RandomSource;
+import net.minecraft.util.Tuple;
 import net.minecraft.world.effect.MobEffect;
 import net.minecraft.world.effect.MobEffectInstance;
 import net.minecraft.world.effect.MobEffects;
@@ -15,51 +17,46 @@ import net.minecraft.world.entity.EquipmentSlot;
 import net.minecraft.world.entity.LivingEntity;
 import net.minecraft.world.item.ItemStack;
 import net.minecraft.world.item.enchantment.Enchantment;
+import net.minecraft.world.item.enchantment.EnchantmentHelper;
 import net.minecraft.world.level.Level;
 import net.minecraft.world.level.block.CropBlock;
 import net.minecraft.world.phys.Vec3;
+import net.tslat.effectslib.api.util.EnchantmentUtil;
 import org.ecorous.dwarventreasures.DwarvenTreasures;
+import org.ecorous.dwarventreasures.server.DwarvenTreasuresServerConfig;
+import org.ecorous.dwarventreasures.world.level.storage.DwarvenTreasuresEntityComponents;
+import org.ecorous.dwarventreasures.world.level.storage.RingComponent;
 
 import java.util.List;
 
 public class DwarvenTreasuresEnchantments
 {
-	public static final String SUSURRUS_TICKS_TAG = "SusurrusTicks";
-
 	public static final Enchantment SUSURRUS = register("susurrus", new RingEnchantment()
 	{
 		@Override
-		public void tick(LivingEntity entity, ItemStack stack)
+		public void tick(LivingEntity livingEntity, ItemStack stack)
 		{
-			super.tick(entity, stack);
-			if (entity.getActiveEffects().stream().noneMatch(effect -> effect.getEffect() == MobEffects.INVISIBILITY && effect.getAmplifier() >= 0 && !effect.endsWithin(20)))
+			super.tick(livingEntity, stack);
+			if (TrinketsApi.getTrinketComponent(livingEntity).orElseThrow().isEquipped(s -> EnchantmentHelper.getItemEnchantmentLevel(this, s) > 0) || livingEntity.getMainHandItem() == stack || livingEntity.getOffhandItem() == stack)
 			{
-				entity.addEffect(new MobEffectInstance(MobEffects.INVISIBILITY, 60));
-			}
-			int ticks = stack.getOrCreateTag().getInt(SUSURRUS_TICKS_TAG);
-			System.out.println(ticks);
-			stack.getOrCreateTag().putInt(SUSURRUS_TICKS_TAG, ticks + 1);
-			List<MobEffect> effects = List.of(MobEffects.MOVEMENT_SLOWDOWN, MobEffects.BLINDNESS, MobEffects.WITHER);
-			for (int i = 1; i <= effects.size(); i++)
-			{
-				int finalI = i;
-				if (ticks >= FIVE_MINUTES * i && entity.getActiveEffects().stream().noneMatch((effect) -> effect.getEffect() == effects.get(finalI - 1) && effect.getAmplifier() >= 0 && !effect.endsWithin(20)))
+				if (livingEntity.getActiveEffects().stream().noneMatch(effect -> effect.getEffect() == MobEffects.INVISIBILITY && effect.getAmplifier() >= 0 && !effect.endsWithin(20)))
 				{
-					entity.addEffect(new MobEffectInstance(effects.get(i - 1), 60));
+					livingEntity.addEffect(new MobEffectInstance(MobEffects.INVISIBILITY, 60));
+				}
+				RingComponent ringComponent = DwarvenTreasuresEntityComponents.RING.get(livingEntity);
+				int ticks = ringComponent.getSusurrusCooldown();
+				System.out.println(ticks);
+				if (!livingEntity.level().isClientSide) ringComponent.increaseSusurrusCooldown();
+				List<MobEffect> effects = List.of(MobEffects.MOVEMENT_SLOWDOWN, MobEffects.BLINDNESS, MobEffects.WITHER);
+				for (int i = 1; i <= effects.size(); i++)
+				{
+					int finalI = i;
+					if (ticks >= FIVE_MINUTES * i && livingEntity.getActiveEffects().stream().noneMatch((effect) -> effect.getEffect() == effects.get(finalI - 1) && effect.getAmplifier() >= 0 && !effect.endsWithin(20)))
+					{
+						livingEntity.addEffect(new MobEffectInstance(effects.get(i - 1), 60));
+					}
 				}
 			}
-		}
-
-		@Override
-		public void onUnequip(LivingEntity entity, EquipmentSlot slot, ItemStack from, ItemStack to, int level)
-		{
-			super.onUnequip(entity, slot, from, to, level);
-		}
-
-		@Override
-		public void onEquip(LivingEntity entity, EquipmentSlot slot, ItemStack from, ItemStack to, int level)
-		{
-			super.onEquip(entity, slot, from, to, level);
 		}
 	});
 	public static final Enchantment ANOINTMENT = register("anointment", new RingEnchantment());
@@ -69,8 +66,11 @@ public class DwarvenTreasuresEnchantments
 		public void tick(LivingEntity livingEntity, ItemStack stack)
 		{
 			super.tick(livingEntity, stack);
-			List<Entity> entities = livingEntity.level().getEntities(livingEntity, livingEntity.getBoundingBox().inflate(8), target -> target.getType().is(EntityTypeTags.UNDEAD));
-			entities.forEach(target -> target.setSecondsOnFire(5));
+			if (TrinketsApi.getTrinketComponent(livingEntity).orElseThrow().isEquipped(s -> EnchantmentHelper.getItemEnchantmentLevel(this, s) > 0) || livingEntity.getMainHandItem() == stack || livingEntity.getOffhandItem() == stack)
+			{
+				List<Entity> entities = livingEntity.level().getEntities(livingEntity, livingEntity.getBoundingBox().inflate(8), target -> DwarvenTreasuresServerConfig.ringEffectTarget.test(livingEntity, target));
+				entities.forEach(target -> target.setSecondsOnFire(5));
+			}
 		}
 	});
 	public static final Enchantment FROST = register("frost", new RingEnchantment()
@@ -79,37 +79,44 @@ public class DwarvenTreasuresEnchantments
 		public void tick(LivingEntity livingEntity, ItemStack stack)
 		{
 			super.tick(livingEntity, stack);
-			Level level = livingEntity.level();
-			List<Entity> entities = level.getEntities(livingEntity, livingEntity.getBoundingBox().inflate(8), entity -> true);
-			for (Entity entity : entities)
+			if (TrinketsApi.getTrinketComponent(livingEntity).orElseThrow().isEquipped(s -> EnchantmentHelper.getItemEnchantmentLevel(this, s) > 0) || livingEntity.getMainHandItem() == stack || livingEntity.getOffhandItem() == stack)
 			{
-				entity.makeStuckInBlock(level.getBlockState(entity.blockPosition()), new Vec3(0.9f, 1.5, 0.9f));
-				if (level.isClientSide) {
-					boolean bl;
-					RandomSource randomSource = level.getRandom();
-					bl = entity.xOld != entity.getX() || entity.zOld != entity.getZ();
-					if (bl && randomSource.nextBoolean()) {
-						level.addParticle(ParticleTypes.SNOWFLAKE, entity.getX(), entity.getY() + 1, entity.getZ(), Mth.randomBetween(randomSource, -1.0f, 1.0f) * 0.083333336f, 0.05f, Mth.randomBetween(randomSource, -1.0f, 1.0f) * 0.083333336f);
+				Level level = livingEntity.level();
+				List<Entity> entities = level.getEntities(livingEntity, livingEntity.getBoundingBox().inflate(8), target -> DwarvenTreasuresServerConfig.ringEffectTarget.test(livingEntity, target));
+				for (Entity entity : entities)
+				{
+					entity.makeStuckInBlock(level.getBlockState(entity.blockPosition()), new Vec3(0.9f, 1.5, 0.9f));
+					if (level.isClientSide)
+					{
+						boolean bl;
+						RandomSource randomSource = level.getRandom();
+						bl = entity.xOld != entity.getX() || entity.zOld != entity.getZ();
+						if (bl && randomSource.nextBoolean())
+						{
+							level.addParticle(ParticleTypes.SNOWFLAKE, entity.getX(), entity.getY() + 1, entity.getZ(), Mth.randomBetween(randomSource, -1.0f, 1.0f) * 0.083333336f, 0.05f, Mth.randomBetween(randomSource, -1.0f, 1.0f) * 0.083333336f);
+						}
 					}
+					entity.setIsInPowderSnow(true);
+					entity.setTicksFrozen(entity.getTicksFrozen() + 20);
+					entity.setRemainingFireTicks(-1);
 				}
-				entity.setIsInPowderSnow(true);
-				entity.setTicksFrozen(entity.getTicksFrozen() + 20);
-				entity.setRemainingFireTicks(-1);
+				List<BlockPos> posList = BlockPos.betweenClosedStream(livingEntity.getBoundingBox().inflate(16)).toList();
+				posList.forEach(pos ->
+				{
+					if (level.getBlockState(pos).getBlock() instanceof CropBlock) level.destroyBlock(pos, true);
+				});
 			}
-			List<BlockPos> posList = BlockPos.betweenClosedStream(livingEntity.getBoundingBox().inflate(16)).toList();
-			posList.forEach(pos ->
-			{
-				if (level.getBlockState(pos).getBlock() instanceof CropBlock)
-					level.destroyBlock(pos, true);
-			});
 		}
 
 		@Override
 		public void onUnequip(LivingEntity entity, EquipmentSlot slot, ItemStack from, ItemStack to, int level)
 		{
 			super.onUnequip(entity, slot, from, to, level);
-			entity.setTicksFrozen(0);
-			entity.setIsInPowderSnow(false);
+			entity.level().getEntities(entity, entity.getBoundingBox().inflate(48)).forEach(entity1 ->
+			{
+				entity1.setTicksFrozen(0);
+				entity1.setIsInPowderSnow(false);
+			});
 		}
 	});
 
@@ -117,4 +124,6 @@ public class DwarvenTreasuresEnchantments
 	{
 		return Registry.register(BuiltInRegistries.ENCHANTMENT, DwarvenTreasures.modLoc(name), enchantment);
 	}
+
+	public static void init() {}
 }
